veyron/lib/stats: Let Glob return inner nodes
With this change, Glob will include inner nodes that don't have a value
when includeValues is false. This means that Glob can be used to
traverse the tree one level at a time. WatchGlob will return all nodes
on the first cycle, and then only the nodes that change afterwards.

Change-Id: I221b4f4672992b3173810c7e6a4b63985eaba1b1