veyron2/vdl/codegen/java: Make all VDL types implement android.os.Parcelable.
This is needed to pass VDL objects within Android efficiently.

Also added type Any, which for now is a JSON-encoded VDL type.
We need Any to be able to communicate with Identity Store from Android,
as the store returns "any".  In the long term, "any" will just work from Java.

Change-Id: I888f929db02273873cfe319af0c8c99b642d1023