veyron/lib/stats/counter: Fix delta value
Before the counter history is full, use the last value as delta. This
means that for the first minute, Value() and Delta1m() will be equal,
for the first 10 minutes, Value() and Delta10m() will be equal, and the
same for 1 hour.

This is more intuitive than the previous behavior, which always
subtracted the oldest value from the newest one.

Change-Id: Ia154f073416e084d4780b8ecd00e49562113524f