veyron/tools/playground: Add response caching to the playground.
Uses an LRU cache (golang/groupcache/lru). Dependent on this CL:
https://veyron-review.googlesource.com/#/c/4253/

Also implements a max size of 65k for both the program and response
size.

NOTE: This is the same CL as https://veyron-review.googlesource.com/#/c/4251/.

That CL was merged before the groupcache dependency made it into
third-party. It broke the build, and I reverted it.

Now that the groupcache dependency is in third-party
(https://veyron-review.googlesource.com/#/c/4253/), this CL should be
good-to-go again.

Change-Id: Ieeab07a49bddbca4e8243e3633828f0361a30ada