veyron/runtimes/google/ipc: Disallow multiple calls to Finish.
I had a bug where my code was calling Finish multiple times.
That resulted in an obscure error in the ipc implementation as
it tried to read the rpc results from the stream a second time.
Of course, there were no results to be read, but the error made
it look like the server had failed to send the results.  This
change makes it clear that multiple calls to Finish are not
allowed.

Change-Id: I42c3aeac5f54c619c7de0102e90ef98fe2096c59