veyron/tools/playground: Fix GOPATH so the vdl compiler can use imports.
The GOPATH did not include $VEYRON_ROOT/veyron/go. As a result, the vdl
compiler could not compile vdl files with imports.

This CL adds $VEYRON_ROOT/veyron/go to the GOPATH in the docker instance
and in the test.sh. It also adds a veyron import to the testdata vdl to
make sure this works.

Change-Id: I8f358a0b32d09c62966b6f30d10c63cf51359380