veyron/services/mgmt/debug: Invoke Glob on the leaves directly.
Instead of sending a local RPC to Glob the leaves, use the invoker
directly. This is more correct and more efficient, especially when the
local endpoint is not actually local, i.e. goes through a proxy.

Also change the Glob method of the logdir invoker to return directories
so that log directories can be traversed one level at a time.

Change-Id: I9d1ff10caa93233a7c234b648fced5630791b670