veyron2/security: Make ThirdPartyRequirements a VDL type.
Motivation: For inter-language interoperability, the wire representation
of a ThirdPartyCaveat and a Discharge need to be specified in VDL. This requires
that ThirdPartyRequirements (a component of ThirdPartyCaveat) be specified in VDL.

While at it, also changed DischargeImpetus to report the blessings held by
the server, as seen by the client, instead of shipping the whole PublicID
(i.e., blessings with cryptographic proof). This saves space and also allows
the discharger to see the server blessings from the client's perspective.

Change-Id: I7797b97d824e64f27bdcbcb260a44a7a4dbf6a18