{veyron,scripts}: removing the REPO_ROOT logic from veyron build and test scripts
Prior to this CL, our Go build wrapper checked if it is being invoked
form within a git repository. If that was the case, the wrapper used
approximate (and currently incorrect) logic to compute the GOPATH,
shadowing a git repository in VEYRON_ROOT with the current
repository. This feature was needed for old Jenkins setup; the new
setup does not need it and this CL removes it. As a consequence, any
developers that keep multiple copies of the same repository of
VEYRON_ROOT will need to provide their own means to building code in
copies that do not have the default path (e.g. $VEYRON_ROOT/veyron).

In addition, this code removes the computation (and use) of REPO_ROOT
from a number of build and test scripts so that these scripts can be
run from outside of a git repository.

Change-Id: Ife19f8bc172bc19e809f566e6597fb4bf9b2c446