{veyron,veyron2}/security: Hack to get old identities working with new.
This is very hacky to allow VEYRON_IDENTITY files created before the
move to the veyron.io package-space to work with binaries built after.
We choose to live with this hack in order to keep backward compatibility
until we switch to the new security model later this month
(and thus incur backward incompatibility once).

Change-Id: I694aa9f48d769e2af26c3cd094747b2b559f131b