veyron2/verror: Rename NotFound and NotAuthorized, introduce NoExistOrNoAccess.
Rename NotFound to NoExist and NotAuthorized to NoAccess.
Introduce a new error, NoExistOrNoAccess, that indicates that a principal may
not access an entity because it does not exist.
NoExist now indicates that a principal may not access an entity either
because the entity does not exist, or the principal is not authorized to
access the entity.

Change-Id: I74116a5fcc7c2486ae356070c63a8c2b7edcb4bd