veyron/services/wsprd: Remove the signature call from the main websocket goroutine.
The call was in the main go routine because we need to know the type of
the stream before we could put in in the outstandingStream map.  The
stream itself could be added later, but the type could, so the fix was
to change the asynchronous stream code to allow delaying the
initializing of both the stream and the type of the stream.

This change also fixes a long standing issue that sending on the server
stream blocked the main pipe thread.

Change-Id: I2e1dd8312d698da7470f888275337edd3bcbed19