veyron/services/mgmt/node/impl/*: add a suidhelper
This is the first patch in a sequence to support running Veyron applications
as the UNIX user associated with the invoking Veyron principal. In this
change, I add the suidhelper code and the node manager's ability to invoke
the helper. A future change will enable the actual setuid functionality
where possible.

Change-Id: I3a2ec6ca1f69c28dc506f5af8c1687335a2fd2f6