veyron/runtimes/google/ipc: Add __debug dispatcher
With this change, requests for object names that begin with a special
keyword ("__debug") are intercepted and redirected to the debug
dispatcher.

If server s is mounted at a/b, a request to a/b/__debug will be routed
to the debug dispatcher. The __debug keyword is only recognized when it
is at the beginning of the object name on a given server.

Change-Id: Id333adc8cc5483d2605b70949235cd00ee3465e5