veyron/runtimes/google/ipc: crude fixes for https://code.google.com/p/envyor/issues/detail?id=257 Addresses issues 1 and 2 in https://docs.google.com/a/google.com/document/d/1C0kxfYhuOcStdV7tnLZELZpUhfQCZj47B0JrzbE29h8/edit See doc for details.
As mentioned in the comment left around the fixes, the fixes are meant to be
short-term, in order to get our builds green. Rather than simply removing the
calls to closing the flow and to removing the vc from the vif from the two call
sites, we should figure out precisely under what circumstances this should be
done.

Change-Id: Ifc1bd5b9ffc97963a585014f13a04be19ae94ce4