veyron/services/mgmt/suidhelper: record username
The node manager tests do not run the suid helper with root
permissions. Consequently, the tests cannot run test apps as different
users. Instead, this change modifies the suid helper to work around
this issue. When in test mode, the suid helper records the username
requested by the node manager and reports this to the test harness
for validation.

Change-Id: Ibedd26b461216ac677fe44774e53862435c9e800