veyron2/ipc,veyron/runtimes/google/ipc: Support for the newer security model.
This commit makes the IPC model (and the corresponding implementation in
veyron/runtimes/google/ipc) respect the new security model (of Principal and
Blessings). All tests are updated to use only the new model (i.e., not use
PrivateID and PublicID).

This commit does break some features, but the author's assessment is that these
features are not widely used enough yet to warrant concern about backward
compatibility. For example: IPCs where clients currently grant blessings to
servers will not work (since the ipc.Request message has changed), and
ThirdPartyCaveat usage with the old model will not function (discharges will
not be fetched by the IPC stack).

(The only known client of the granted blessings was the nodemanager Claim
method, tests for which have been temporarily disabled)

Future steps (towards removing the old model completely):
(1) Update the veyron/runtimes/google/rt implementation so that it
    sets up IPC clients and servers with the new primitives
(2) Update the identityd server and identity tool to use the new
    primitives
(3) Replace application code using the old primitives with new.

Change-Id: Ie9a9d3b3dba5ef8d9a6fc22b5f6bc09ee3f02cbe