veyron/services/mgmt: Fix broken tests with a hack.
veyron/services/mgmt/node/impl started failing the TestAppLifeCycle test after
https://veyron-review.googlesource.com/#/c/5164/ was submitted.

Debugging traced this failure to the innocent change in the packages
being imported in veyron/services/mgmt/node/impl/impl_test.go and the
fact that veyron/lib/testutil calls flag.Parse in its init() method.

The aforementioned change changed the package dependencies of
veyron/services/mgmt/suidhelper/impl. This, in turn, caused the compiler to
change the package initialization order such that veyron/lib/testutil was being
initailized before veyron/services/mgmt/suidhelper/impl.
As a result of this reordering, the flags defined in the impl package were not
available as the flag parsing happened before these new flags were defined.

This change introduces a hack to ensure the correct ordering by adding a
dependency from veyron/lib/testutil to veyron/services/mgmt/suidhelper/flag.

However, this is just a hack (and the fact that the test ever had
a working initialization order was a stroke of luck). The correct fix
is to ensure that flag.Parse it not called in the init of any package
(other than "main").

The correct fix is more involved, so in order to get the build
green again, I propose this temporary hack.

Change-Id: I30cdc14d0d749c205bc9c17b221b213a2df3f8f4