"veyron/services/identity": BlessUsingAccessToken returns email
Modified BlessUsingAccessToken to along with the blessing also
return the email for which the blessing was generated. The email
provides clients a means to tag the blessing (e.g. WSPR would
use the email while adding the Blessing to the identity mananger).

Additionally, also updated the implementation to use the new
security model when possible.

Change-Id: Iab180ff412769a507f8117a4eaf375cce0829a9a