"veyron/services/wsrd/principal": Principal Manager
This CL adds a PrincipalManager to WSPR to manage root accounts
and create principals for various apps/origins. The behavior is
mostly the same as the existing IDManager, with the following
difference:

* Account names are human-readable blessing strings. They are meant
for referring to the account in UIs and are mapped to the account's
Blessings object in the manager. The Blessings object is also required
to have the account name as a blessing but this is not an enforced
constraint.

* The AccountsMatching method returns Accounts by matching the provided
pattern with the names of the accounts (as opposed to the Blessings objects
associated with the account)

Additonally, this CL also copies the JSPublicIDHandles from identity
as JSBlessingsHandles to this package.

Change-Id: I250413d81fde4bfe80160ba4152801c3e198fe76