veyron: cleanup of shell test scripts making sure that non-zero exit codes are not ignored and informative error messages are generated
This CL also fixes a process leak that was cause by using 'exec' in
the agent shell test.

This CL depends on go/vcl/5585 and go/vcl/5591 and should not be
submitted before the other two CLs are. Until the other two CLs are
submitted, the presubmit check of shell test is expected to fail.

Change-Id: I5a77ecc5592378e8f62bcce1f9ac5487affa9705