"veyron/services/wsprd": Flag to enable new model
This CL adds a flag to wsprd to indicate that the new
security model should be used. This flag will be used
by the JS tests to test against the new model. The flag
is meant to go away once the old security model is
killed.

Additionally the CL also makes the following two changes:
1) Send RemoteBlessing strings as well in the
ServerRPCRequestContext.
2) Fix a bug in how the new security model is enabled
for the wspr Controller.

Change-Id: Ic0ebf3ef97b14fa4acf0d87372a0f055d72f3703