veyron/runtimes/google/ipc: Fix flaky TestReconnect
There is a race between when the server dies and when the VIF gets the
error. If the next StartCall happens before the VIF is shutdown, the
client see an EOF error instead of the expect failure to connect.

A simple way to fix this is to also accept EOF as valid error.

Change-Id: I304586b3941cec1921bfe5435965366a104909f7