"veyron/tools/principal": Overwrite flag for create
At the moment, the create command on the principal tools
complains if it is provided with a directly that already
contains a principal object. This causes some inconvenience
in tests and tutorials.

This CL resolves the issues by adding a --overwrite flag to
the principal that overwrites any existing principal data in
the directory.

Change-Id: I08ed5abb89c4a2d74de42586846b49df73a6f0f4