veyron/runtimes/google/naming: Bugfix in error propagation and test setup.
(1) The glob implementation would not extract the "application error"
(error returned by the server's implementation of the Glob method)
but instead only return framework errors. Also, it was not calling
ipc.Call.Finish, which is not strictly necessary but considered good
practice.

(2) The authorization policy in the test was the "default" one, which
should fail when the client and server are two independent principals.
It just happened to work on in the test because of the unfortunate
use of a global "trusted keys" map in the old security model. However,
this will be going away with the new security model and the
"allow everyone" policy change in this test seems to reflect intent
more accurately.

Change-Id: I630e00ba6f77000e75dd1c5b3a53ae288edd1419