"veyron/services/wsprd/wspr_nacl": Initialize principal from key specified in the config.
The WSPR_NACL instance currents expects a veyron identity to be passed
to it via a ppapi.Var. This has two problems (1) it uses the old
security model, and (2) it expects the identity from the config which
makes it difficult to have different identities per instance.

This CL resolves (1) by obtaining a private key from the config and then
creating an in-memory principal that is passed to NewWSPR as a
RuntimePrincipal option.

To resolve (2), we would have to manage the key within WSPR, possibly
persisting it in some Chrome storage that is directly accessible via a NACL
API. This is left as a TODO.

See also: https://veyron-review.googlesource.com/#/c/5992/

Change-Id: I01ade443bed308e3ea1853d6e7451bdab6faebc6