veyron/lib/textutil: Add textutil package, including LineWriter.
LineWriter implements an io.Writer filter that formats input text
into output lines with a given target width in runes.  It
performs word-wrapping.

In order to implement this as an io.Writer filter, I also added
UTF8ChunkDecoder which decodes a stream of UTF-8 encoded runes
that may be arbitrarily chunked.  There's also a RuneChunkDecoder
interface to allow us to easily add UTF-16 support in the future.

I've updated the cmdline package to use this new support, and
updated the vdl tool docs with the new formatting to give an idea
of the changes.  I'll re-generate all docs in a subsequent CL.

Change-Id: Ia83cffe71d6fd43a0806fe028ff92d0d4ff079c3