veyron2/vdl: Add typeobject to parser, compiler and go codegen.
A large part of of this CL is renaming from typeval->typeobject.

Note that some support for typeobject already existed; the easy
stuff like supporting the typeobject type in the compiler and
codegen.  This CL fills in the holes, adding support for
typeobject constants, and making some minor changes to make this
easier.

I'm particularly pleased with the one weird trick in
codegen/golang/const.go; I'd originally put off adding go codegen
support for typeobject consts because I thought it was lots of
mindless work, but it turns out it's really easy.

Change-Id: I2fdab9d6582ea83c17851d75ffe845bc505e5136