veyron/runtimes/google/naming/namespace:
This is the second to last CL to get rid of //.  Except for
all the tests that look for and add //'s they should no
longer be needed. ns.Mount, ns.Unmount, etc now add the
NotResolve flag to their StartCalls where needed.  The
next CL will update the zillion+1 tests and eradicate
all code to handle //.

1) Add ns.ResolveX and ns.ResolveToMountTableX that
return a naming.MountEntry instead of a simple list of
names.  Reimplement ns.Resolve and ns.ResolveToMountTable
as calls to these.

2) Make name resolution obey the ServesMountTable bit
in the MountEntry and the one in the Endpoint.   This is
a better way to stop than putting a // somewhere in the
middle of the name.

3) Change the time in naming.MountEntry to be an
absolute expiration time. It used to be a duration
which was kind of meaningless except at the exact
moment it was returned.

Change-Id: Ifa50ac0be3adc0040f3ab525299a706e5e60cec4