veyron/runtimes/google/ipc: Explicitly implement default authorization.
The default authorization can be expressed in terms of ACLs, however
some folks had expressed the desire to explicitly pull out the default
policy so that:
(a) It can be more succinctly expressed
    (the language used in documentation is clearly visible in a 5 line
    code snippet)
(b) Error messages are more meaningful (does wonders in the tutorial).
    This comes at the cost of the server revealing the authorization
    policy it is using, but that seems okay.

Furthermore, there seems to be growing dissent when it comes to
using Labels in ACLs, so decoupling the default policy from ACLs
seems like the right thing to do.

Change-Id: Idec4fa9d6d9ce932bd61f8c718b67c294135f7b1