veyron2/security: Make the timestamp and method tags available in security.Context
This commit makes two additions to the Context interface:
(1) Timestamp: The time at which the context was created/the time at
which the authorization check is to be made.
The expiry caveat now uses the time from the Context (instead of
time.Now()) to validate. This allows for more deterministic caveat
validation (eg. blessings.ForContext(ctx) will always return the
same result for the same 'ctx' now. Without this change, it could
return different results at different times based on the time
difference between the two calls).

(2) MethodTags: This is motivated by:
https://docs.google.com/a/google.com/document/d/1DZUu2sGKrf-b4p9JFdIbN0ULhz9Loa8reuPNKsdMKD0/edit?usp=sharing

Change-Id: I7f979948670c09ae9bdd483cce4bff839624f158