veyron/services/mgmt/binary: HTTP download for binary service.
The implementation uses the http file server functionality, and we provide
an implementation of http.File that wraps the content parts we have in the
binary server.

For now, make any content available via http (DownloadURL is still not
implemented).  We'll revisit that later.

As part of this change, factor out some code from dispatcher and invoker
that was also needed for the http implementation.

While at it, fix some issues discovered while working with this codebase.
Among them, how Upload/Download handle invalid part indices.

Change-Id: Ief4153b68930d5edbf8b83e36c3b1396369d5f3c