veyron/services/mgmt/binary/binaryd: Make port configurable.
Instead of being hardcoded, the port is now settable by a flag.  In addition,
we print out the HTTP server's ip:port (after trying to figure out a suitable
ip address to publish).  This will be useful generally, but in particular it
allows us to fix test.sh by letting the port be dynamically generated and then
scraped from the log (change to test.sh to come in a separate CL).

Change-Id: Ia443ea55396d117208efa1f935b16bcd25a1a1d5