veyron/services/mgmt/node/impl: identity management for child apps (agent-less).
Implementation of principal creation/blessing for apps in the node manager.
This version does not integrate with the agent, but rather uses the local
filesystem to create and store the principal, and passes it to the app using
VeyronCredentials env.

The app principal is created upon instance creation (during Start). The
principal is blessed with (1) the blessing granted by the called of Start and
(2) a blessing derived from the node manager's default blessing (so that the
node manager can talk to the app's cycle manager; and the app can talk back to
the node manager).

Because the node manager now provides apps with credentials, we don't need to
'hack' identities into the app's envelopes in the unit test.

Change-Id: Ifea29e9177729c9fa0dba9353d90c58066278024