veyron/runtimes/google/ipc: Rewrite the ipc.Server Glob implementation
This is the 3rd of a series of changes that add a simple
interface for objects to participate in the namespace.

With this change, the ipc.Server now intercepts all Glob (and __Glob)
requests and uses the new VGlobber interface to construct the proper
response.

Subsequent changes will update the namespace library to use __Glob, and
update some of the existing Glob implementations to use the new
interface.

Change-Id: Iec65d78af21fbb8fbcdc4625c4c709a9d9e915ce