veyron/runtimes/google/ipc: Switch everything to __Glob.
This is the 4th of a series of changes that add a simple
interface for objects to participate in the namespace.

With this change, the ipc Server no longer intercepts to the Glob
method. It now only intercepts __Glob.

The only left after this is to change (some of) the existing Glob
implementations to the simpler interface.

Change-Id: I0b9cbebc781e9793ef62f54e01da0f37847670af