TBR Browspr - WSPR for the browser
This is a version of WSPR that is build around postMessage rather than HTTP/Websocket requests.

A number of parts of WSPR have been refactored.

In the future, we may want WSPR to use the same interface as browspr (it can use the callbacks used for postmessage), just to consolidate code.

Change-Id: I338265a4305ba74c4c46e47051aeb850b2219582